chore: extract duration helper for migrations
